Add TLS support
In order to add TLS support, add a gopher-global-tls-mode. That's
because the *gopher* buffer loses its gopher-tls-mode setting after a
reload. Use open-network-stream instead of make-network-process and
add sentinel, filter and coding system manually instead of passing it
along as parameters. This requires a small change in how gopher-coding
is structured. Use the opportunity and make UTF-8 the default.

Remove gopher-extra-network-args.

gopher-sentinel-directory-listing now splits on \n instead of \r\n.